+++ tests/modules/test_defaultAggregator.py
@@ -1,6 +1,6 @@
 """ test """
-import time
 import logging
+import time
 
 from bzt.modules.aggregator import ResultsReader, DataPoint, KPISet
 from tests import BZTestCase, r, rc, err
@@ -12,12 +12,14 @@ class TestDefaultAggregator(BZTestCase):
         super(TestDefaultAggregator, self).setUp()
         self.obj = ResultsReader()
         self.obj.track_percentiles = [25, 50, 75, 80, 90, 95, 99, 99.9, 100]
+        self.obj.buffer_scale_idx = str(float(self.obj.track_percentiles[-1]))
         self.obj.buffer_len = 1
 
     def test_1(self):
         obj = self.obj
 
         mock = MockReader()
+        mock.buffer_scale_idx = '100.0'
         mock.data.append((1, "", 1, r(), r(), r(), 200, None, ''))
         mock.data.append((2, "", 1, r(), r(), r(), 200, None, ''))
         mock.data.append((2, "", 1, r(), r(), r(), 200, None, ''))
@@ -48,7 +50,7 @@ class TestDefaultAggregator(BZTestCase):
         obj = self.obj
 
         mock = MockReader()
-
+        mock.buffer_scale_idx = '100.0'
         obj.add_listener(mock)
 
         res = {}
@@ -71,3 +73,50 @@ class TestDefaultAggregator(BZTestCase):
 
         for point in mock.datapoints(True):
             pass
+
+    def test_0buffer_scaling(self):
+        obj = self.obj
+
+        mock = MockReader()
+        obj.add_listener(mock)
+
+        # t_stamp, label, conc, r_time, con_time, latency, r_code, error, trname
+        mock.min_buffer_len = 1
+        mock.buffer_len = 1
+        mock.buffer_scale_idx = '50.0'
+
+        buffer_len = mock.buffer_len
+        for i in range(5):
+            mock.data.append((100 + i, "", 1, 2, 2, 2, 200, None, ''))
+        points = list(mock.datapoints())
+        points = list(mock.datapoints())
+        self.assertTrue(mock.buffer_len > buffer_len)
+        buffer_len = mock.buffer_len
+
+        for i in range(10):
+            mock.data.append((200 + i, "", 1, 3, 3, 3, 200, None, ''))
+        points = list(mock.datapoints())
+        points = list(mock.datapoints())
+        self.assertTrue(mock.buffer_len > buffer_len)
+        buffer_len = mock.buffer_len
+
+        for i in range(20):
+            mock.data.append((300 + i, "", 1, 4, 4, 4, 200, None, ''))
+        points = list(mock.datapoints())
+        points = list(mock.datapoints())
+        self.assertTrue(mock.buffer_len > buffer_len)
+        buffer_len = mock.buffer_len
+
+        for i in range(15):
+            mock.data.append((400 + i, "", 1, 1, 1, 1, 200, None, ''))
+        points = list(mock.datapoints())
+        points = list(mock.datapoints())
+        self.assertTrue(mock.buffer_len < buffer_len)
+        buffer_len = mock.buffer_len
+
+        for i in range(30):
+            mock.data.append((500 + i, "", 1, 1, 1, 1, 200, None, ''))
+        points = list(mock.datapoints())
+        points = list(mock.datapoints())
+        self.assertTrue(mock.buffer_len < buffer_len)
+        pass