+++ virtualenv-test.bash
@@ -7,7 +7,7 @@ source build/bin/activate
 # install depends
 pip install --upgrade colorlog pyyaml psutil lxml cssselect nose urwid coverage six pylint selenium progressbar33 locustio pyvirtualdisplay
 
-pylint -d R0903,R0904,C0301,C0302,C0111,E1103,R0201,R0902,W0511,F0401,E0611,R0801,R0913,W0613 -f parseable bzt > build/pylint.out || echo
+pylint -d R0903,R0904,C0301,C0302,C0111,E1103,R0201,R0902,W0511,F0401,E0611,R0801,R0913,W0613,C0412 -f parseable bzt > build/pylint.out || echo
 
 # run unit tests
 ./run-test.sh