+++ bzt/modules/provisioning.py
@@ -16,6 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+import datetime
 import time
 
 from bzt.engine import Provisioning
@@ -26,6 +27,31 @@ class Local(Provisioning):
     """
     Local provisioning means we start all the tools locally
     """
+    def _get_start_shift(self, shift):
+        if shift == '':
+            return 0
+
+        time_formats = ['%Y-%m-%d %H:%M:%S',
+                        '%Y-%m-%d %H:%M',
+                        '%H:%M:%S',
+                        '%H:%M']
+
+        for time_format in time_formats:
+            try:
+                date = datetime.datetime.strptime(shift, time_format)
+            except ValueError:
+                continue
+            except TypeError:
+                self.log.warning('Start time must be string type ("%s"), ignored', time_format[0])
+                break
+            today = datetime.date.today()
+            if today > date.date():
+                date = datetime.datetime(today.year, today.month, today.day, date.hour, date.minute, date.second)
+            return time.mktime(date.timetuple()) - self.start_time
+        else:
+            self.log.warning('Unrecognized time format: %s ("%s" required), ignored', shift, time_formats[0])
+
+        return 0
 
     def prepare(self):
         """
@@ -36,10 +62,14 @@ class Local(Provisioning):
             self.log.debug("Preparing executor: %s", executor)
             executor.prepare()
             self.engine.prepared.append(executor)
-            executor.delay = dehumanize_time(executor.execution.get('delay', '0'))
 
     def startup(self):
         self.start_time = time.time()
+        for executor in self.executors:
+            start_shift = self._get_start_shift(executor.execution.get('start-at', ''))
+            delay = dehumanize_time(executor.execution.get('delay', 0))
+            executor.delay = delay + start_shift
+            self.log.debug("Delay setup: %s(start-at) + %s(delay) = %s", start_shift, delay, executor.delay)
 
     def _start_modules(self):
         for executor in self.executors: