+++ bzt/cli.py
@@ -137,7 +137,7 @@ class CLI(object):
             if self.options.no_system_configs is None:
                 self.options.no_system_configs = False
 
-            self.engine.configure(configs, not self.options.no_system_configs)
+            merged_config = self.engine.configure(configs, not self.options.no_system_configs)
 
             # apply aliases
             for alias in self.options.aliases:
@@ -149,6 +149,8 @@ class CLI(object):
             if self.options.option:
                 overrider = ConfigOverrider(self.log)
                 overrider.apply_overrides(self.options.option, self.engine.config)
+
+            self.engine.create_artifacts_dir(configs, merged_config)
 
             self.engine.prepare()
             self.engine.run()