+++ bzt/__init__.py
@@ -17,7 +17,7 @@ import os
 import sys
 from abc import abstractmethod
 
-VERSION = "1.3.3"
+VERSION = "1.4.0"
 
 
 class RCProvider(object):