+++ site/dat/docs/Changelog.md
@@ -6,7 +6,8 @@
  - do not create hostaliases file when aliases are not specified
  - add `force-parent-sample` option to JMeter executor
  - add `compile-target-java` option for Selenium
- - add dynamic buffer scaling ability to ResultsReader  
+ - add dynamic buffer scaling ability to ResultsReader
+ - add scheduling ability with `start-at` parameter  
  - apply overrides and cli-aliases before creating artifacts-dir
 
 ## 1.3.3 <sup>24 mar 2016</sup>