+++ site/dat/docs/Changelog.md
@@ -8,6 +8,7 @@
  - add `compile-target-java` option for Selenium
  - add dynamic buffer scaling ability to ResultsReader
  - add scheduling ability with `start-at` parameter  
+ - apply overrides and cli-aliases before creating artifacts-dir
 
 ## 1.3.3 <sup>24 mar 2016</sup>
  - add new `hostaliases` setting for all executors